Move UiState/UiEvent from HomeScreen to HomePresenter
