+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/HomeScreen.kt
@@ -1,24 +1,7 @@
 package org.cru.godtools.ui.dashboard.home
 
-import com.slack.circuit.runtime.CircuitUiEvent
-import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.screen.Screen
 import kotlinx.parcelize.Parcelize
-import org.cru.godtools.ui.banner.BannerType
-import org.cru.godtools.ui.tools.ToolCard
 
 @Parcelize
-data object HomeScreen : Screen {
-    data class UiState(
-        val dataLoaded: Boolean = true,
-        val banner: BannerType? = null,
-        val spotlightLessons: List<ToolCard.State> = emptyList(),
-        val favoriteTools: List<ToolCard.State> = emptyList(),
-        val eventSink: (UiEvent) -> Unit = {},
-    ) : CircuitUiState
-
-    sealed interface UiEvent : CircuitUiEvent {
-        data object ViewAllFavorites : UiEvent
-        data object ViewAllTools : UiEvent
-    }
-}
+data object HomeScreen : Screen
